Split production.c into at_host and sensors.
This split makes more sense since AT commands can be
used at any time now. Remove the term production.
