+++ prj.conf
@@ -12,14 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-CONFIG_BT_DEVICE_NAME="C209-AoA-Tag"
+CONFIG_BT_DEVICE_NAME="u-blox C209 DF Tag"
 CONFIG_ASSERT=y
 CONFIG_BT_ID_MAX=1
 CONFIG_BT_CTLR_TX_PWR_PLUS_4=y
 CONFIG_BT_CTLR_TX_PWR_DYNAMIC_CONTROL=y
 
 CONFIG_BT=y
-CONFIG_BT_DEVICE_NAME="u-blox C209 DF Tag"
 CONFIG_BT_MAX_CONN=2
 
 CONFIG_BT_EXT_ADV=y
@@ -55,6 +54,7 @@ CONFIG_CLOCK_CONTROL_NRF_K32SRC_RC=y
 
 CONFIG_MAIN_STACK_SIZE=4096
 CONFIG_SYSTEM_WORKQUEUE_STACK_SIZE=4096
+CONFIG_BT_RX_STACK_SIZE=2048
 
 CONFIG_NVS=y
 CONFIG_FLASH=y
@@ -76,6 +76,9 @@ CONFIG_PM=y
 CONFIG_PM_DEVICE=y
 CONFIG_PM_DEVICE_RUNTIME=y
 CONFIG_REBOOT=y
+
+CONFIG_SEND_SENSOR_DATA_IN_PER_ADV_DATA=y
+CONFIG_ALLOW_REMOTE_AT_OVER_NUS=y
 
 # Build for use with Open Bootloader which is preflashed on NINA-B40X modules
 # CONFIG_FLASH_BASE_ADDRESS=0x1000
